01_Семестр/Python - advanced level/Lab Work/06_case/django-blog-main/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.1
django-crispy-forms==2.3
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
